0.2.0.alpha.2-SNAPSHOT - fixes, trying to reduce boilerplate code regarding the SerialName of id/version/username fields
