+++ src/main/kotlin/com/kamikazejam/datastore/base/async/handler/crud/AsyncUpdateHandler.kt
@@ -6,7 +6,7 @@ import com.kamikazejam.datastore.base.async.handler.AsyncHandler
 import com.kamikazejam.datastore.base.async.result.DefiniteResult
 import com.kamikazejam.datastore.base.async.result.Failure
 import com.kamikazejam.datastore.base.async.result.Success
-import com.kamikazejam.datastore.mode.store.Store
+import com.kamikazejam.datastore.store.Store
 
 @Suppress("unused")
 class AsyncUpdateHandler<K : Any, X : Store<X, K>>(
