+++ src/main/kotlin/com/kamikazejam/datastore/base/extensions/CollectionExtensions.kt
@@ -9,7 +9,7 @@ import com.kamikazejam.datastore.base.async.handler.crud.AsyncUpdateHandler
 import com.kamikazejam.datastore.base.async.result.Empty
 import com.kamikazejam.datastore.base.async.result.Failure
 import com.kamikazejam.datastore.base.async.result.Success
-import com.kamikazejam.datastore.mode.store.Store
+import com.kamikazejam.datastore.store.Store
 import org.jetbrains.annotations.NonBlocking
 
 /**
