+++ src/main/kotlin/com/kamikazejam/datastore/base/loader/StoreLoader.kt
@@ -1,7 +1,7 @@
 package com.kamikazejam.datastore.base.loader
 
 import com.kamikazejam.datastore.base.Collection
-import com.kamikazejam.datastore.mode.store.Store
+import com.kamikazejam.datastore.store.Store
 
 /**
  * This class is responsible for loading a [Store] when requested from a [Collection].
