+++ src/main/kotlin/com/kamikazejam/datastore/base/storage/StorageLocal.kt
@@ -1,7 +1,7 @@
 package com.kamikazejam.datastore.base.storage
 
 import com.kamikazejam.datastore.base.Collection
-import com.kamikazejam.datastore.mode.store.Store
+import com.kamikazejam.datastore.store.Store
 import com.kamikazejam.datastore.util.DataStoreFileLogger
 import java.util.concurrent.ConcurrentHashMap
 import java.util.concurrent.ConcurrentMap
