+++ src/main/kotlin/com/kamikazejam/datastore/connections/storage/mongo/TransactionResult.kt
@@ -1,5 +1,5 @@
 package com.kamikazejam.datastore.connections.storage.mongo
 
-import com.kamikazejam.datastore.mode.store.Store
+import com.kamikazejam.datastore.store.Store
 
 internal class TransactionResult<K : Any, X : Store<X, K>>(val store: X?, val dbStore: X? = null)
