+++ src/main/kotlin/com/kamikazejam/datastore/store/object/StoreObjectCollection.kt
@@ -1,48 +1,47 @@
-package com.kamikazejam.datastore.mode.profile
+package com.kamikazejam.datastore.store.`object`
 
 import com.google.common.base.Preconditions
 import com.kamikazejam.datastore.DataStoreRegistration
 import com.kamikazejam.datastore.DataStoreSource
 import com.kamikazejam.datastore.base.StoreCollection
 import com.kamikazejam.datastore.base.async.handler.crud.AsyncCreateHandler
-import com.kamikazejam.datastore.base.extensions.read
 import com.kamikazejam.datastore.base.log.CollectionLoggerService
 import com.kamikazejam.datastore.base.log.LoggerService
-import com.kamikazejam.datastore.api.event.StoreProfileQuitEvent
-import com.kamikazejam.datastore.mode.profile.storage.ProfileStorageDatabase
-import com.kamikazejam.datastore.mode.profile.storage.ProfileStorageLocal
-import com.kamikazejam.datastore.mode.store.StoreProfile
-import com.kamikazejam.datastore.util.PlayerUtil
-import com.mongodb.DuplicateKeyException
+import com.kamikazejam.datastore.store.`object`.storage.ObjectStorageDatabase
+import com.kamikazejam.datastore.store.`object`.storage.ObjectStorageLocal
+import com.kamikazejam.datastore.store.StoreObject
+import com.mongodb.*
 import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.flow
+import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.runBlocking
 import org.bukkit.Bukkit
-import org.bukkit.entity.Player
 import java.util.*
 import java.util.concurrent.ConcurrentHashMap
 import java.util.concurrent.ConcurrentMap
 
 @Suppress("unused")
-abstract class StoreProfileCollection<X : StoreProfile<X>> @JvmOverloads constructor(
+abstract class StoreObjectCollection<X : StoreObject<X>> @JvmOverloads constructor(
     module: DataStoreRegistration,
-    override val instantiator: (UUID, Long, String?) -> X,
+    override val instantiator: (String, Long) -> X,
     name: String,
     storeClass: Class<X>,
-    logger: (StoreCollection<UUID, X>) -> LoggerService = { store -> CollectionLoggerService(store) },
+    logger: (StoreCollection<String, X>) -> LoggerService = { store -> CollectionLoggerService(store) },
 ) :
-    StoreCollection<UUID, X>(name, UUID::class.java, storeClass, module, logger),
-    ProfileCollection<X> {
-    private val loaders: ConcurrentMap<UUID, StoreProfileLoader<X>> = ConcurrentHashMap()
-    override val localStore: ProfileStorageLocal<X> = ProfileStorageLocal()
-    override val databaseStore: ProfileStorageDatabase<X> by lazy { ProfileStorageDatabase(this) }
-    override val defaultInitializer: (X) -> X = { it }
+    StoreCollection<String, X>(name, String::class.java, storeClass, module, logger),
+    ObjectCollection<X>
+{
+    private val loaders: ConcurrentMap<String, StoreObjectLoader<X>> = ConcurrentHashMap()
+    override val localStore: ObjectStorageLocal<X> = ObjectStorageLocal()
+
+    override val databaseStore: ObjectStorageDatabase<X> by lazy { ObjectStorageDatabase(this) }
 
     init {
         // Start this collection
         val success = runBlocking { start() }
         if (!success) {
             // Data loss is not tolerated in DataStore, shutdown to prevent issues
-            DataStoreSource.get().logger.severe("Failed to start Profile Collection: $name")
+            DataStoreSource.get().logger.severe("Failed to start Object Cache: $name")
             Bukkit.shutdown()
         }
     }
@@ -73,16 +72,16 @@ abstract class StoreProfileCollection<X : StoreProfile<X>> @JvmOverloads constru
     // ----------------------------------------------------- //
     @Suppress("DuplicatedCode")
     @Throws(DuplicateKeyException::class)
-    override fun create(key: UUID, initializer: (X) -> X): AsyncCreateHandler<UUID, X> {
+    override fun create(key: String, initializer: (X) -> X): AsyncCreateHandler<String, X> {
         Preconditions.checkNotNull(initializer, "Initializer cannot be null")
 
         return AsyncCreateHandler(this) {
             try {
                 // Create a new instance in modifiable state
-                val initial: X = instantiator(key, 0L, null)
+                val initial: X = instantiator(key, 0L)
                 initial.initialize(this)
 
-                // Initialize the store
+                // Initialize the store (make sure the id and version were not changed)
                 val store: X = initializer(initial)
                 assert(store.id == key) { "Store ID must match key on Creation!" }
                 assert(store.version == 0L) { "Store version must be 0 on Creation!" }
@@ -102,38 +101,60 @@ abstract class StoreProfileCollection<X : StoreProfile<X>> @JvmOverloads constru
         }
     }
 
-    override suspend fun readAllFromDatabase(cacheStores: Boolean): Flow<X> {
-        return databaseStore.getAll()
+    @Throws(DuplicateKeyException::class)
+    override fun create(initializer: (X) -> X): AsyncCreateHandler<String, X> {
+        return this.create(UUID.randomUUID().toString(), initializer)
     }
 
     // ----------------------------------------------------- //
-    //                         Cache                         //
+    //                 Misc Collection Methods               //
     // ----------------------------------------------------- //
-    override fun loader(key: UUID): StoreProfileLoader<X> {
+    override fun loader(key: String): StoreObjectLoader<X> {
         Preconditions.checkNotNull(key)
-        return loaders.computeIfAbsent(key) { s: UUID -> StoreProfileLoader(this, s) }
+        return loaders.computeIfAbsent(key) { s: String -> StoreObjectLoader(this, s) }
     }
 
-    override fun keyToString(key: UUID): String {
-        return key.toString()
+    override fun keyToString(key: String): String {
+        return key
     }
 
-    override fun keyFromString(key: String): UUID {
-        Preconditions.checkNotNull(key)
-        return UUID.fromString(key)
+    override fun keyFromString(key: String): String {
+        return key
+    }
+
+    override suspend fun readAllFromDatabase(cacheStores: Boolean): Flow<X> = flow {
+        // Create an Iterable that iterates through all database objects, and updates local objects as necessary
+        databaseStore.getAll(this@StoreObjectCollection).map { dbStore: X ->
+            // Load the local object
+            val local: X? = localStore.get(dbStore.id)
+
+            // If we want to cache, and have a local store that's newer -> update the local store
+            // Note, if not caching then we won't update any local stores and won't cache the db store
+            val dbVer = dbStore.version
+            val localVer = local?.version ?: 0
+            if (cacheStores && local != null && dbVer >= localVer) {
+                this@StoreObjectCollection.updateStoreFromNewer(local, dbStore)
+            }
+
+            // Find the store object to return
+            val ret = local ?: dbStore
+            // Verify it has the correct cache and cache it if necessary
+            ret.initialize(this@StoreObjectCollection)
+            ret
+        }
     }
 
     override val cached: Collection<X>
         get() = localStore.localStorage.values
 
-    override fun readFromCache(key: UUID): X? {
+    override fun readFromCache(key: String): X? {
         return localStore.get(key)
     }
 
-    override suspend fun readFromDatabase(key: UUID, cacheStore: Boolean): X? {
+    override suspend fun readFromDatabase(key: String, cacheStore: Boolean): X? {
         val o: X? = databaseStore.get(key)
         if (cacheStore) {
-            o?.let { store -> this.cache(store) }
+            o?.let { this.cache(it) }
         }
         return o
     }
@@ -141,39 +162,7 @@ abstract class StoreProfileCollection<X : StoreProfile<X>> @JvmOverloads constru
     override val localCacheSize: Long
         get() = localStore.size()
 
-    override suspend fun getIDs(): Flow<UUID> {
+    override suspend fun getIDs(): Flow<String> {
         return databaseStore.getKeys()
-    }
-
-    override suspend fun getOnline(): Collection<X> {
-        return Bukkit.getOnlinePlayers()
-            .filter { PlayerUtil.isFullyValidPlayer(it) }
-            .mapNotNull { player ->
-                this.read(player).await().getOrNull()
-            }
-    }
-
-    override fun readFromCache(player: Player): X? {
-        Preconditions.checkNotNull(player)
-        return readFromCache(player.uniqueId)
-    }
-
-    override suspend fun readFromDatabase(player: Player, cacheStore: Boolean): X? {
-        Preconditions.checkNotNull(player)
-        val o: X? = databaseStore.get(player.uniqueId)
-        if (cacheStore) {
-            o?.let { store -> this.cache(store) }
-        }
-        return o
-    }
-
-    override fun removeLoader(uuid: UUID) {
-        Preconditions.checkNotNull(uuid)
-        loaders.remove(uuid)
-    }
-
-    override fun onProfileLeaving(player: Player, profile: X) {
-        val event = StoreProfileQuitEvent(player, this, profile)
-        Bukkit.getPluginManager().callEvent(event)
     }
 }
