+++ src/main/kotlin/com/kamikazejam/datastore/store/object/storage/ObjectStorageDatabase.kt
@@ -1,8 +1,8 @@
-package com.kamikazejam.datastore.mode.`object`.storage
+package com.kamikazejam.datastore.store.`object`.storage
 
 import com.kamikazejam.datastore.base.Collection
 import com.kamikazejam.datastore.base.storage.StorageDatabase
-import com.kamikazejam.datastore.mode.store.StoreObject
+import com.kamikazejam.datastore.store.StoreObject
 
 @Suppress("unused")
 class ObjectStorageDatabase<X : StoreObject<X>>(collection: Collection<String, X>) :
