+++ src/main/kotlin/com/kamikazejam/datastore/store/object/storage/ObjectStorageLocal.kt
@@ -1,7 +1,7 @@
-package com.kamikazejam.datastore.mode.`object`.storage
+package com.kamikazejam.datastore.store.`object`.storage
 
 import com.kamikazejam.datastore.base.storage.StorageLocal
-import com.kamikazejam.datastore.mode.store.StoreObject
+import com.kamikazejam.datastore.store.StoreObject
 
 @Suppress("unused")
 class ObjectStorageLocal<X : StoreObject<X>> : StorageLocal<String, X>() {
