+++ src/main/kotlin/com/kamikazejam/datastore/store/profile/StoreProfileLoader.kt
@@ -1,11 +1,11 @@
-package com.kamikazejam.datastore.mode.profile
+package com.kamikazejam.datastore.store.profile
 
 import com.google.common.base.Preconditions
 import com.kamikazejam.datastore.DataStoreSource
 import com.kamikazejam.datastore.base.extensions.update
 import com.kamikazejam.datastore.base.loader.StoreLoader
-import com.kamikazejam.datastore.mode.profile.listener.ProfileListener
-import com.kamikazejam.datastore.mode.store.StoreProfile
+import com.kamikazejam.datastore.store.profile.listener.ProfileListener
+import com.kamikazejam.datastore.store.StoreProfile
 import com.kamikazejam.datastore.util.Color
 import com.kamikazejam.datastore.util.DataStoreFileLogger
 import org.bukkit.ChatColor
