+++ src/main/kotlin/com/kamikazejam/datastore/store/profile/storage/ProfileStorageDatabase.kt
@@ -1,8 +1,8 @@
-package com.kamikazejam.datastore.mode.profile.storage
+package com.kamikazejam.datastore.store.profile.storage
 
 import com.kamikazejam.datastore.base.Collection
 import com.kamikazejam.datastore.base.storage.StorageDatabase
-import com.kamikazejam.datastore.mode.store.StoreProfile
+import com.kamikazejam.datastore.store.StoreProfile
 import java.util.UUID
 
 class ProfileStorageDatabase<X : StoreProfile<X>>(collection: Collection<UUID, X>) :
